services/ai_component/pyproject.toml
[project]
name = "ai_component"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium[classic-control,mujoco]>=0.29.1",
    "matplotlib>=3.9.2",
    "mujoco>=3.2.3",
    "numpy>=2.1.1",
    "pyyaml>=6.0.2",
    "torch>=2.4.1",
]
